requirement.txt
rembg
pillow
numpy
matplotlib